requirements.txt
Django==5.0.1
django-crispy-forms==2.1
Pillow==10.2.0
django-notifications-rest==1.0.1
django-ckeditor==6.7.0
python-dateutil==2.8.2